Turn MSON attributes into Elements
This commit removes “attributes” properties of Resource / Action /
Payload object so they can be used in the future by the Element
structure.

The `Attributes` object is replaced by `Data Structure` object which
will be eventually defined in MSON AST as an element of `dataStructure`
type.

In addition this commit brings some clarification to deprecation
messages.